commit 7: eliminado parámetro de listEmployee. Cambiado el parámetro de setSalary a float

+++ AccesoDatos/src/act06/Employee.java
@@ -126,7 +126,7 @@ public class Employee implements Serializable {
 		return salary;
 	}
 
-	public void setSalary(int salary) {
+	public void setSalary(float salary) {
 		if (salary >= 0) {
 			this.salary = salary;
 		} else {
